// frontend/src/api/model/memoryGraphDto.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * backend | v1
 * OpenAPI spec version: 1.0.0
 */
import type { MemoryGraphLink } from './memoryGraphLink';
import type { MemoryGraphNode } from './memoryGraphNode';

export interface MemoryGraphDto {
    nodes: MemoryGraphNode[];
    links: MemoryGraphLink[];
}